JH comments
